+++ tests/test_workflow_config.py
@@ -4,7 +4,8 @@ from pathlib import Path
 
 import pytest
 
-from myteam.workflow.config import ProjectWorkflowDefaults, load_project_workflow_defaults
+from myteam.workflow.config import load_project_workflow_defaults
+from myteam.workflow.models import ProjectWorkflowDefaults
 
 
 def test_load_project_workflow_defaults_returns_none_when_missing(tmp_path: Path):
